Change method to get account ID and way the S3 bucket is created
The way we get the AWS Account ID forces this to be performed by an IAM user, the better option is to move to STS's get-caller-identity API instead. This allows the operator of the install script to use roles if required.

The way we generate buckets should be regional, rather than global.